Check for reference assemblies and AlwaysBuild
We now check the T4 files if it has a referenced
assembly. If it has a referenced assembly, we
check if the referenced assembly is newer than the
generated file. if it's newer, we regenerate the file.

I also added a new property called always run. Setting
this to true will always run the custom tools regardless
if the referenced assembly is new or not as long as
RunCustomToolOnBuild is set to true.